+++ hutool-core/src/main/java/com/xiaoleilu/hutool/util/PageUtil.java
@@ -7,7 +7,7 @@ package com.xiaoleilu.hutool.util;
  * 
  */
 public class PageUtil {
-	
+
 	/**
 	 * 将页数和每页条目数转换为开始位置和结束位置<br>
 	 * 此方法用于不包括结束位置的分页方法<br>
@@ -47,5 +47,64 @@ public class PageUtil {
 			return 0;
 		}
 		return totalCount % numPerPage == 0 ? (totalCount / numPerPage) : (totalCount / numPerPage + 1);
+	}
+
+	/**
+	 * 分页彩虹算法<br>
+	 * 来自：https://github.com/iceroot/iceroot/blob/master/src/main/java/com/icexxx/util/IceUtil.java<br>
+	 * 通过传入的信息，生成一个分页列表显示
+	 * 
+	 * @param currentPage 当前页
+	 * @param pageCount 总页数
+	 * @param displayCount 每屏展示的页数
+	 * @return 分页条
+	 */
+	public static int[] rainbow(int currentPage, int pageCount, int displayCount) {
+		boolean isEven = true;
+		isEven = displayCount % 2 == 0;
+		int left = displayCount / 2;
+		int right = displayCount / 2;
+
+		int length = displayCount;
+		if (isEven) {
+			right++;
+		}
+		if (pageCount < displayCount) {
+			length = pageCount;
+		}
+		int[] result = new int[length];
+		if (pageCount >= displayCount) {
+			if (currentPage <= left) {
+				for (int i = 0; i < result.length; i++) {
+					result[i] = i + 1;
+				}
+			} else if (currentPage > pageCount - right) {
+				for (int i = 0; i < result.length; i++) {
+					result[i] = i + pageCount - displayCount + 1;
+				}
+			} else {
+				for (int i = 0; i < result.length; i++) {
+					result[i] = i + currentPage - left + (isEven ? 1 : 0);
+				}
+			}
+		} else {
+			for (int i = 0; i < result.length; i++) {
+				result[i] = i + 1;
+			}
+		}
+		return result;
+
+	}
+
+	/**
+	 * 分页彩虹算法(默认展示10页)<br>
+	 * 来自：https://github.com/iceroot/iceroot/blob/master/src/main/java/com/icexxx/util/IceUtil.java
+	 * 
+	 * @param currentPage 当前页
+	 * @param pageCount 总页数
+	 * @return 分页条
+	 */
+	public static int[] rainbow(int currentPage, int pageCount) {
+		return rainbow(currentPage, pageCount, 10);
 	}
 }